# doc/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store Daily QA System API",
    "description": "API for store staff daily QA system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/survey/answers": {
      "post": {
        "tags": ["survey"],
        "summary": "Save Survey Answer",
        "description": "サーベイ回答保存",
        "operationId": "save_survey_answer_api_survey_answers_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SurveyAnswerRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Survey answer saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyOperationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "get": {
        "tags": ["survey"],
        "summary": "Get Survey Answers",
        "description": "サーベイ回答履歴取得",
        "operationId": "get_survey_answers_api_survey_answers_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ページ番号",
              "default": 1,
              "title": "Page"
            },
            "description": "ページ番号"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "1ページあたりの件数",
              "default": 30,
              "title": "Limit"
            },
            "description": "1ページあたりの件数"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey answers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyAnswerListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/survey/answers/{survey_id}": {
      "put": {
        "tags": ["survey"],
        "summary": "Update Survey Answer",
        "description": "サーベイ回答更新",
        "operationId": "update_survey_answer_api_survey_answers__survey_id__put",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Survey Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyAnswerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Survey answer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyOperationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Survey not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "get": {
        "tags": ["survey"],
        "summary": "Get Survey Answer",
        "description": "サーベイ回答詳細取得",
        "operationId": "get_survey_answer_api_survey_answers__survey_id__get",
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Survey Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey answer retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyAnswerResponse"
                }
              }
            }
          },
          "404": {
            "description": "Survey not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorResponse" }
        },
        "type": "object",
        "required": ["error"],
        "title": "APIError",
        "description": "API エラーレスポンス"
      },
...
      "AnswerData": {
        "properties": {
          "question_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Question Id",
            "description": "質問ID"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text",
            "description": "質問内容"
          },
          "question_type": { "$ref": "#/components/schemas/QuestionType" },
          "answer_value": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Answer Value",
            "description": "回答値"
          },
          "options": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Options",
            "description": "選択肢（choice/ratingタイプの場合）"
          }
        },
        "type": "object",
        "required": ["question_id", "question_text", "question_type"],
        "title": "AnswerData",
        "description": "回答データ"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "エラーコード"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "エラーメッセージ"
          },
          "details": {
            "anyOf": [{}, { "type": "null" }],
            "title": "Details",
            "description": "エラー詳細"
          }
        },
        "type": "object",
        "required": ["code", "message"],
        "title": "ErrorResponse",
        "description": "エラーレスポンス"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": ["text", "choice", "rating", "number"],
        "title": "QuestionType",
        "description": "質問タイプ"
      },
...
      "SurveyAnswerListItem": {
        "properties": {
          "survey_id": {
            "type": "string",
            "title": "Survey Id",
            "description": "サーベイID"
          },
          "user_cd": {
            "type": "string",
            "title": "User Cd",
            "description": "ユーザーコード"
          },
          "str_cd": {
            "type": "string",
            "title": "Str Cd",
            "description": "店舗コード"
          },
          "answer_count": {
            "type": "integer",
            "title": "Answer Count",
            "description": "回答数"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          }
        },
        "type": "object",
        "required": [
          "survey_id",
          "user_cd",
          "str_cd",
          "answer_count",
          "created_at",
          "updated_at"
        ],
        "title": "SurveyAnswerListItem",
        "description": "サーベイ回答リスト項目"
      },
      "SurveyAnswerListResponse": {
        "properties": {
          "answers": {
            "items": { "$ref": "#/components/schemas/SurveyAnswerListItem" },
            "type": "array",
            "title": "Answers",
            "description": "サーベイリスト"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "総件数"
          },
          "current_page": {
            "type": "integer",
            "title": "Current Page",
            "description": "現在のページ番号"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "総ページ数"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "1ページあたりの件数"
          }
        },
        "type": "object",
        "required": [
          "answers",
          "total_count",
          "current_page",
          "total_pages",
          "limit"
        ],
        "title": "SurveyAnswerListResponse",
        "description": "サーベイ回答履歴レスポンス"
      },
      "SurveyAnswerRequest": {
        "properties": {
          "survey_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Survey Id",
            "description": "サーベイID"
          },
          "answers": {
            "items": { "$ref": "#/components/schemas/AnswerData" },
            "type": "array",
            "title": "Answers",
            "description": "回答データリスト"
          }
        },
        "type": "object",
        "required": ["survey_id", "answers"],
        "title": "SurveyAnswerRequest",
        "description": "サーベイ回答保存リクエスト"
      },
      "SurveyAnswerResponse": {
        "properties": {
          "survey_id": {
            "type": "string",
            "title": "Survey Id",
            "description": "サーベイID"
          },
          "user_cd": {
            "type": "string",
            "title": "User Cd",
            "description": "ユーザーコード"
          },
          "str_cd": {
            "type": "string",
            "title": "Str Cd",
            "description": "店舗コード"
          },
          "answers": {
            "items": { "$ref": "#/components/schemas/AnswerData" },
            "type": "array",
            "title": "Answers",
            "description": "回答データリスト"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          }
        },
        "type": "object",
        "required": [
          "survey_id",
          "user_cd",
          "str_cd",
          "answers",
          "created_at",
          "updated_at"
        ],
        "title": "SurveyAnswerResponse",
        "description": "サーベイ回答レスポンス"
      },
      "SurveyAnswerUpdateRequest": {
        "properties": {
          "answers": {
            "anyOf": [
              {
                "items": { "$ref": "#/components/schemas/AnswerData" },
                "type": "array"
              },
              { "type": "null" }
            ],
            "title": "Answers",
            "description": "回答データリスト"
          }
        },
        "type": "object",
        "title": "SurveyAnswerUpdateRequest",
        "description": "サーベイ回答更新リクエスト"
      },
      "SurveyOperationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作結果メッセージ"
          },
          "survey_id": {
            "type": "string",
            "title": "Survey Id",
            "description": "サーベイID"
          }
        },
        "type": "object",
        "required": ["message", "survey_id"],
        "title": "SurveyOperationResponse",
        "description": "サーベイ操作レスポンス"
      }
    }
  }
}